+++ src/cqml/keys.py
@@ -6,8 +6,11 @@ DB_NAME='nauto'
 DATE_COL='daily_ts'
 DATE_UNIQ='serial_number'
 
+TRACE=True
+
 cAlias='|'
 kAbove='above'
+kAny='any'
 kArgs='args'
 kCast='as_number'
 kCols='cols'